application_progress.py
#!/usr/bin/env python3
"""
Twitter API申請進行状況追跡
"""
import os
from datetime import datetime, timedelta
from dotenv import load_dotenv

def track_application_progress():
    """申請進行状況の追跡"""
    print("📋 Twitter API申請進行状況\n")
    
    # 申請チェックリスト
    checklist = [
        ("X Developer Portal アクセス", "https://developer.twitter.com/en/portal/dashboard"),
        ("基本情報入力", "名前、国、メールアドレス"),
        ("用途選択", "Academic research を選択"),
        ("プロジェクト説明", "STEP_BY_STEP_APPLICATION.md 参照"),
        ("データ利用方針", "教育目的、プライバシー配慮を明記"),
        ("申請送信", "内容確認後に送信"),
        ("承認待ち", "1-7営業日"),
        ("Bearer Token取得", "承認後にDeveloper Portalで取得"),
        ("環境設定", ".envファイルにトークン設定"),
        ("動作確認", "python api_status_tracker.py")
    ]
    
    print("✅ 申請チェックリスト:")
    print("-" * 50)
    for i, (step, detail) in enumerate(checklist, 1):
        print(f"{i:2d}. {step}")
        print(f"    → {detail}")
        print()
    
    # 現在の状況確認
    load_dotenv()
    bearer_token = os.getenv("TWITTER_BEARER_TOKEN")
    
    if bearer_token and bearer_token != "your_actual_bearer_token_here":
        print("🎉 申請完了・設定済み!")
        print(f"Bearer Token: {bearer_token[:20]}...")
        return "completed"
    else:
        print("⏳ 申請進行中または未開始")
        
        # 申請状況の推定
        print("\n📊 推定申請状況:")
        print("□ 申請未開始 または")
        print("□ 申請済み・承認待ち または") 
        print("□ 承認済み・設定未完了")
        
        return "in_progress"

def show_immediate_actions():
    """今すぐできるアクション"""
    print("\n" + "="*50)
    print("🎯 今すぐできるアクション")
    print("="*50)
    
    print("\n1. 申請開始 (5-10分)")
    print("   → https://developer.twitter.com/en/portal/dashboard")
    print("   → STEP_BY_STEP_APPLICATION.md を参照")
    
    print("\n2. 申請待ちの間の開発継続")
    print("   → データベース設計・実装")
    print("   → マッチングアルゴリズム開発")
    print("   → UI/UX改善")
    
    print("\n3. 申請状況確認")
    print("   → Developer Portal でステータス確認")
    print("   → メール通知の確認")

def estimate_timeline():
    """申請タイムライン予測"""
    print("\n📅 申請タイムライン予測")
    print("-" * 30)
    
    today = datetime.now()
    
    # 申請日（今日）
    print(f"申請日: {today.strftime('%Y-%m-%d')} (今日)")
    
    # 承認予定日（3-7日後）
    approval_min = today + timedelta(days=3)
    approval_max = today + timedelta(days=7)
    print(f"承認予定: {approval_min.strftime('%Y-%m-%d')} - {approval_max.strftime('%Y-%m-%d')}")
    
    # 開発完了予定日（承認後1-2日）
    completion = approval_max + timedelta(days=2)
    print(f"実装完了: {completion.strftime('%Y-%m-%d')} 頃")
    
    print(f"\n⏰ 実データでのテスト開始まで: 約{(completion - today).days}日")

def main():
    """メイン実行"""
    print(f"🕐 {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    print("="*60)
    
    status = track_application_progress()
    
    if status == "in_progress":
        show_immediate_actions()
        estimate_timeline()
        
        print("\n💡 申請のコツ:")
        print("   - 教育目的であることを強調")
        print("   - プライバシー配慮を明記")
        print("   - 具体的な社会課題解決を説明")
        
    print(f"\n📚 参考資料:")
    print("   - STEP_BY_STEP_APPLICATION.md (詳細手順)")
    print("   - API_APPLICATION_GUIDE.md (申請ガイド)")
    print("   - application_template.txt (記入テンプレート)")

if __name__ == "__main__":
    main()