requirements.txt
flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
pandas==2.1.4
scikit-learn==1.3.2
tweepy==4.14.0
gunicorn==21.2.0
Werkzeug==3.0.6